+++ website/routes/doc.js
@@ -2,7 +2,7 @@ var express = require('express');
 var path    = require('path');
 var app     = express();
 
-//app.use(express.static('doc'));
-app.use('/', express.static(path.join(__dirname, '../../doc')));
+app.use(express.static('doc'));
+//app.use('/', express.static(path.join(__dirname, './doc')));
 
 module.exports = app;